Extracted code into separate ES2015 modules.
